% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row.R
\name{get_row}
\alias{get_row}
\title{Work out a particular row of a metrics table.}
\usage{
get_row(row, data, skip = 0)
}
\arguments{
\item{row}{The row of the table to be analysed.}

\item{data}{The data that contains the row to convert.}

\item{skip}{Should a particular row be skipped? If so, enter the number or vector of numbers.}
}
\description{
Work out a particular row of a metrics table.
}
