% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chapter_metrics_tables.R
\name{make_chapter_metrics_tables}
\alias{make_chapter_metrics_tables}
\title{Make a individual chapter's metrics tables}
\arguments{
\item{data}{The data you want to convert into a metrics table.
Note that this should be downloaded from the Google Doc (see 'get_metrics_data()')}

\item{useLines}{The vector of bulk lines from the main metrics table.}
}
\description{
Make a individual chapter's metrics tables
}
