% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics_table.R
\name{make_metrics_table}
\alias{make_metrics_table}
\title{Make a metrics table for Grattan}
\arguments{
\item{data}{The data you want to convert into a metrics table.
Note that it should be in the format of a metrics table.}
}
\description{
Make a metrics table for Grattan
}
